options-dashboard/pyproject.toml
[project]
name = "medha-options-dashboard"
version = "0.1.0"
description = "Streamlit options signal dashboard (Dhan API only)"
requires-python = ">=3.12"
dependencies = [
    "dhanhq>=2.0.2",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
